refactor: move web3Handler logic to loadContracts function
